Updated documentation for CMakeSelectCwd (#138)
Updated documentation to match the new behavior of the CMakeSelectCwd command. The docs now reference the single argument given version as well.

new select_cwd command with arfs

The argument given is the absolute path of the directory

CMakeSelectCwd with parameters

Using CMakeSelectCwd with one parameter will result in the plugin using that parameter as the cwd.